% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcrp-get.R
\name{bcrp_get}
\alias{bcrp_get}
\title{Get data BCRP by link}
\usage{
bcrp_get(.link = NULL)
}
\arguments{
\item{.link}{Url from https://estadisticas.bcrp.gob.pe/estadisticas/series/}
}
\value{
A list with data from BCRP links
}
\description{
Create a tibble from BCRP data
}
\details{
Read table from links of BCRP data, and convert into tibbles.
}
\examples{
library(rvest)
library(purrr)
library(PeruData)

down_load <- c(
 "https://estadisticas.bcrp.gob.pe/estadisticas/series/mensuales/resultados/PN01270PM/html"
, "https://estadisticas.bcrp.gob.pe/estadisticas/series/mensuales/resultados/PN01308PM/html"
)
dfs <- bcrp_get(down_load)
dfs
}
\author{
Jhon Flores Rojas
}
