% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_functions.R
\name{clean_sf}
\alias{clean_sf}
\title{clean_sf}
\usage{
clean_sf(.sf, .simplify = T, keep = 0.05)
}
\arguments{
\item{.sf}{A sf object}

\item{.simplify}{Reduce memory usage}

\item{keep}{Size to keep}
}
\value{
A clean sf data
}
\description{
clean_sf
}
\examples{
library(PeruData)
clean_sf(map_peru_dep)
}
