% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_files.R
\name{dir_files}
\alias{dir_files}
\title{dir_files}
\usage{
dir_files(.data, .mod, .mod_len)
}
\arguments{
\item{.data}{create folders to sort data}

\item{.mod}{Module or survey}

\item{.mod_len}{Number of modules or surveys}
}
\value{
Folders created
}
\description{
dir_files
}
\examples{
#dir_files("enaho", c("01", "02"), 2)
}
