% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inei_enaho.R
\name{enaho_clean}
\alias{enaho_clean}
\title{enaho_clean}
\usage{
enaho_clean(.enaho_data)
}
\arguments{
\item{.enaho_data}{ENAHO database}
}
\value{
Standard data cleaning
}
\description{
enaho_clean
}
\examples{
#enaho_2018 <- haven::read_dta("enaho/modulo 01/enaho_2018.dta")
#enaho_2018 |>
#    enaho_clean()
}
