% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enaho_rename}
\alias{enaho_rename}
\title{ENAHO
ENAHO new names}
\format{
An object of class \code{character} of length 6.
}
\usage{
enaho_rename
}
\description{
ENAHO
ENAHO new names
}
\keyword{datasets}
