% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_functions.R
\name{get_centroid}
\alias{get_centroid}
\title{get_centroid}
\usage{
get_centroid(.sf)
}
\arguments{
\item{.sf}{A sf object}
}
\value{
X Y columns with centroid
}
\description{
get_centroid
}
\examples{
library(PeruData)
#map_peru_dist |>
#  filter(depa == "huanuco") |>
#   get_centroid()
}
