% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inei_db}
\alias{inei_db}
\title{INEI diccionario
Diccionario de codigos de INEI}
\format{
An object of class \code{list} of length 4.
}
\usage{
inei_db
}
\description{
INEI diccionario
Diccionario de codigos de INEI
}
\keyword{datasets}
