% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inei_enaho.R
\name{inei_enaho}
\alias{inei_enaho}
\title{inei_enaho}
\usage{
inei_enaho(.mod, .year, .tp = "stata", .trash = c("tabla", "otro", "dic"))
}
\arguments{
\item{.mod}{Surveys (module) to download from the National Household Survey - Peru}

\item{.year}{Years of surveyss to download}

\item{.tp}{File extension \code{c("stata", "spss", "csv")} for years greater than 2019 csv is available}

\item{.trash}{Not relevant files}
}
\value{
A "solo-data" forder with the database sorted
}
\description{
inei_enaho
}
\examples{
library(PeruData)
years <- c("2010", "2019")
mod <- c("01", "03", "05", "85")
# inei_enaho(mod, years)
}
