% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inei_endes.R
\name{inei_endes}
\alias{inei_endes}
\title{Download data from INEI survey ENDES}
\usage{
inei_endes(.mod, .year, .tp = "spss")
}
\arguments{
\item{.tp}{}
}
\value{

}
\description{
Download data from INEI survey ENDES
}
