% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_peru_depa}
\alias{map_peru_depa}
\title{Mapas del peru
map_peru_depa: Departamentos}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 25 rows and 4 columns.
}
\usage{
map_peru_depa
}
\description{
Mapas del peru
map_peru_depa: Departamentos
}
\keyword{datasets}
