% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_peru_dist}
\alias{map_peru_dist}
\title{Mapas del peru
map_peru_dist: Distrital}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1872 rows and 4 columns.
}
\usage{
map_peru_dist
}
\description{
Mapas del peru
map_peru_dist: Distrital
}
\keyword{datasets}
