% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_peru_peru}
\alias{map_peru_peru}
\title{Mapa del peru
map_peru_depa: Peru nacional}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 4 columns.
}
\usage{
map_peru_peru
}
\description{
Mapa del peru
map_peru_depa: Peru nacional
}
\keyword{datasets}
