% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_peru_prov}
\alias{map_peru_prov}
\title{Mapas del peru
map_peru_prov: Provincial}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 196 rows and 6 columns.
}
\usage{
map_peru_prov
}
\description{
Mapas del peru
map_peru_prov: Provincial
}
\keyword{datasets}
