% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_map_classic}
\alias{theme_map_classic}
\title{Theme classic maps}
\usage{
theme_map_classic(
  ...,
  lt = "dashed",
  cl_line = "gray80",
  title_font = "serif",
  title_size = 14,
  caption_font = "serif",
  legend_pst = c(0.9, 0.1)
)
}
\arguments{
\item{...}{adicional elements}

\item{lt}{linetype}

\item{cl_line}{line color}

\item{title_font}{font title}

\item{title_size}{size title}

\item{caption_font}{caption font}

\item{legend_pst}{lenged position}
}
\description{
Theme classic maps
}
\examples{
librarian::shell(tidyverse, sf, PeruData)
ggplot() +
geom_sf(data = map_peru_depa, fill = "gray90", color = "black") +
geom_sf(data = filter(map_peru_depa, depa == "piura"), fill = "steelblue") +
theme_map_classic()
}
\author{
Jhon Flores Rojas
}
