% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_map_sp}
\alias{theme_map_sp}
\title{Theme for location plot (subplot)}
\usage{
theme_map_sp(cl_borde = "black", fill_map = "white", ...)
}
\arguments{
\item{cl_borde}{plot border color}

\item{fill_map}{plot fill color}

\item{...}{adicional elements}
}
\description{
Theme for location plot (subplot)
}
\examples{
librarian::shell(tidyverse, sf, PeruData)
ggplo() +
geom_sf(data = map_peru_depa, fill = "white", color = "black") +
geom_sf(data = filter(map_peru_depa, depa == "piura", fill = "gray20", color = "gray20") +
theme_map_sp()
}
\author{
Jhon Flores Rojas
}
