% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_functions.R
\name{tidy_text}
\alias{tidy_text}
\title{tidy_text}
\usage{
tidy_text(.x)
}
\arguments{
\item{.x}{A vector of type character}
}
\value{
Cleaning data from a vector or columns, removing unnecessary spaces or special characters
}
\description{
tidy_text
}
\examples{
dirty <- c("    a ", "universo  ", "íóñ")
tidy_text(dirty)
}
