% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ubigeo_peru}
\alias{ubigeo_peru}
\title{Ubigeo peru
Codigo de ubigeo peru}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1874 rows and 8 columns.
}
\usage{
ubigeo_peru
}
\description{
Ubigeo peru
Codigo de ubigeo peru
}
\keyword{datasets}
