% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_files.R
\name{dir_files}
\alias{dir_files}
\title{Create INEI folders}
\usage{
dir_files(.data, .mod, .mod_len)
}
\arguments{
\item{.data}{create folders to sort data}

\item{.mod}{Module or survey}

\item{.mod_len}{Number of modules or surveys}
}
\value{
Folder for INEI database
}
\description{
Create folders for INIE database
}
\examples{
dir_files("enaho", c("01", "02"), 2)
}
\author{
Jhon Flores rojas
}
