% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inei_enaho.R
\name{enaho_clean}
\alias{enaho_clean}
\title{Clean datasets from ENAHO}
\usage{
enaho_clean(.enaho_data)
}
\arguments{
\item{.enaho_data}{ENAHO database}
}
\value{
Standard data cleaning
}
\description{
Create a tidy data from ENAHO
}
\details{
Clean data from ENAHO and prepare for a panel data
}
\examples{
# library(tidyverse)
library(PeruData)
library(haven)

years <- c("2010", "2019")
mod <- c("01", "03", "05", "85")
# inei_enaho(mod, years) #not run example


#enaho_2018 <- haven::read_dta(here::here("enaho/modulo 01/enaho_2019.dta"))
#enaho_2018 |>
#    enaho_clean()
}
\author{
Jhon Flores Rojas
}
