% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{igp}
\alias{igp}
\title{IGP data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 22767 rows and 7 columns.
}
\usage{
igp
}
\description{
IGP data
}
\keyword{datasets}
