% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inei_enaho.R
\name{inei_enaho}
\alias{inei_enaho}
\title{Download ENAHO from INEE database}
\usage{
inei_enaho(.mod, .year, .tp = "stata", .trash = c("tabla", "otro", "dic"))
}
\arguments{
\item{.mod}{Surveys (module) to download from the National Household Survey - Peru}

\item{.year}{Years of surveyss to download}

\item{.tp}{File extension \code{c("stata", "spss", "csv")} for years greater than 2019 csv is available}

\item{.trash}{Not relevant files}
}
\value{
A "solo-data" forder with datasets
}
\description{
Download ENAHO datasets from INEI database
}
\details{
Download data and sort into separate folders
}
\examples{
#library(tidyverse)
library(PeruData)
years <- c("2010", "2019")
mod <- c("01", "03", "05", "85")
# inei_enaho(mod, years) #not run example
}
\author{
Jhon Flores Rojas
}
